BUG: GHS3DPlugin: Fixed a bug in the max memory computation
The default maximum memory was defined as a short integer which caused the maximum memory to be badly initialised on computers with big memory.
If the available memory was more than 32768 Mo (2**15 Mo) the max memory was cast from int to short and could be initialised with any value included a negative one. In this case a 20Mo default value was used by GHS3D